{{cookiecutter.project_slug}}/requirements.txt
asyncpg==0.30.0
fastapi==0.115.4
greenlet==3.1.1
icecream==2.1.3
jinja2==3.1.4
jsonschema==4.23.0
pytest-mock==3.14.0
pytest==8.3.3
python-decouple==3.8
{% if cookiecutter.use_keycloak == "y" %}
python-keycloak==4.3.0
{% endif %}
{% if cookiecutter.use_redis == "y" %}
redis==5.0.8
{% endif %}
ruff==0.8.0
sqlalchemy==2.0.36
sqlmodel==0.0.22
typer==0.13.1
uvicorn==0.32.0
websockets==14.1
